==> step1_convert_to_integer.ipynb <==
############################################################################################################################
#############################                     understand dataset                    ##################################
############################################################################################################################

# import pd to analyze data
import pandas as pd


# read data as framework
df = pd.read_csv('./interview_dataset.csv')


# check the framework
df.head()

# read the datatype in each attribute
df.info() 

#### data types are differenct
#### need to convert the string into number for further analysis - to do prediction

############################################################################################################################
#############################                     convert string into int                    ##################################
############################################################################################################################


#############################      convert 'treatment company' into int

#### split the column of 'treatment company'
df[['treatment','company','tc_number']] = df['treatment company'].str.split(expand=True)

#### drop the column of 'treatment company','treatment','company', but keep 'tc_number' as 'treatment company'
df_drop_tc = df.drop(['treatment company','treatment','company'], axis=1)

# df_drop_tc.head()

#############################      convert 'operator' into int

##### split the column of 'operator'
df_drop_tc[['operator_name','op_number']] = df_drop_tc['operator'].str.split(expand=True) 

##### drop the column of ''operator','operator_name', but keep 'op_number' as 'operator'
df_drop_op = df_drop_tc.drop(['operator','operator_name'], axis=1)

# df_drop_op.head()

#############################      convert 'treatment company' and 'operator' into numbers

df_drop_op['tc_number'] = df_drop_op.tc_number.astype(int)
df_drop_op['op_number'] = df_drop_op.op_number.astype(int)

# df_drop_op.info()

#############################      convert 'date on production' into int

#### convert 'date on production' into unix timestamp seconds

df_drop_op['date_sec'] = (pd.to_datetime(df_drop_op['date on production'])- pd.Timestamp("1970-01-01")) // pd.Timedelta('1s')

# df_drop_op.info()
# df_drop_op['date_sec'].head()

#### drop 'date on production',keep 'date_sec' as 'date on production'

df_drop_dp = df_drop_op.drop(['date on production'], axis=1)

#### finished converting the data type 
df_drop_dp.head()

#############################      finish the data type converting by saving it as a new dataframe
df_int = df_drop_dp
#df_int.info()

############################################################################################################################
#############################                     saving dataframe into csv.                    ##################################
############################################################################################################################

#############################   put 'production' to the first coulumn   

#### make a list of all of the columns in the df
cols = list(df_int.columns.values) 

#### remove b from list
cols.pop(cols.index('production')) 

#### create new dataframe with columns in the order you want
df_int = df_int[['production']+cols] 

# df_int.head()

#############################   save data 
df_int.to_csv('data_integer_shift_column.csv',index=False)
