==> step2_find critical variables.ipynb <==
############################################################################################################################
#############################                     convert data into df                    ##################################
############################################################################################################################


import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
from patsy import dmatrices
import statsmodels.api as sm


from statsmodels.stats.outliers_influence import variance_inflation_factor
from statsmodels.tools.tools import add_constant


# read data as framework
df = pd.read_csv('./data_integer_shift_column.csv')

# check the framework
df.head()

############################################################################################################################
###############################                     count missing data            ##########################################
############################################################################################################################

# the percentage of missing data in df
print('-------------------------------')
print('missing percentage (%):')
print('-------------------------------')
df.isna().sum()/len(df)*100

#########################################################################################################################

#As aboserved
    ## ~75% data in 'breakdown pressure'
    ## ~58% data in 'water saturation'
    ## ~15% data in 'well spacing'
    ## ~13% data in 'proppant volume'
    ## ~12% data in 'porosity'
    ## ~8% data in 'isip'
    ## ~5% data in 'azimuth'
    ## ~2% data in 'tvd'
    ## ~2% data in 'toc'
    ## ~2% data in 'youngs modulus'
    
    
#############################################   using KNN to predict the missing data   #####################################################


##############################    scale the data before prediction

from sklearn.preprocessing import MinMaxScaler

scaler = MinMaxScaler()
df_sc = pd.DataFrame(scaler.fit_transform(df), columns = df.columns)
# df_sc.head()


##############################    KNN prediction

from sklearn.impute import KNNImputer

imputer = KNNImputer(n_neighbors=5)
df_sc_knn = pd.DataFrame(imputer.fit_transform(df_sc),columns = df_sc.columns)

df_sc_knn.isna().sum().sum()

############################################################################################################################
#################################################    find critical variables   #############################################
############################################################################################################################


################################################# use all varables to predict, find adj R2

import statsmodels.api as sm

x = df_sc_knn.drop(['production'], axis=1)
y = df_sc_knn['production']
                
X2 = sm.add_constant(x)
est = sm.OLS(y, X2)
est2 = est.fit()
print(est2.summary())

#########################################      delete the attribute with too large P value 

### according to reference, if |P value| < 0.05, the feature is significant
### to be conservative, we choose |P value| < 0.2, and delete the value =>0.2

p_values = est2.summary2().tables[1]['P>|t|']

rstl_p = pd.DataFrame(p_values)

rstl_p.loc[(rstl_p['P>|t|']>=0.2)]

###################################   calculate the variance inflation factor (VIF)   #######################################


## Multicollinearity occurs when two or more predictors in the model are correlated
## it provide redundant information about the response
## Multicollinearity was measured by variance inflation factors (VIF) and tolerance
## VIF >10 is not acceptable

X = add_constant(df_sc_knn)


# VIF dataframe 
vif_data = pd.DataFrame() 
vif_data["feature"] = X.columns 
  
# calculating VIF for each feature 
vif_data["VIF"] = [variance_inflation_factor(X.values, i) 
                          for i in range(len(X.columns))] 

############################## list the attributes with 'VIF'>10

rslt_df = vif_data[vif_data['VIF']>10] 
  
print('\feature with VIF>10 :\n', rslt_df) 

#############  Accroding to P value, we can delete

                # azimuth
                # md (ft)
                # footage lateral length
                # porpoise deviation
                # porpoise count
                # acoustic impedance
                # shale footage
                # water saturation
                # toc
                # vcl
                # proppant fluid ratio
                # tc_number
                
#############  Accroding to VIF value, we can delete

                # md (ft)
                # footage lateral length 
                # youngs modulus
            
################################
# we will delete all attributes recommended by P value and VIF
# except, 'young modulus', as the reference has reported the correlation to production 

############################## they are 'md(ft)', 'footage lateral length', and 'youngs modulus'

# df_sc_drop = df_sc_knn.drop(['md (ft)', 'footage lateral length'], axis=1)

df_sc_drop = df_sc_knn.drop(['azimuth',
                             'md (ft)',
                             'footage lateral length',
                             'porpoise deviation',
                             'porpoise count',
                             'acoustic impedance',
                             'shale footage',
                             'water saturation', 
                             'toc',
                             'vcl',
                             'proppant fluid ratio',
                             'tc_number'], axis=1)

################################################# use remaining varables to predict, find adj R2

x = df_sc_drop.drop(['production'], axis=1)
y = df_sc_drop['production']
                
X2 = sm.add_constant(x)
est = sm.OLS(y, X2)
est2 = est.fit()
print(est2.summary())

#################################
## as observed the adjusted R square is improved by deleting the variables

################################# correlation ##########################################

## the coefficient returns a value between -1 and 1 
## the value represents the limits of correlation from a full negative correlation to a full positive correlation
## according to reference, |value| <0.1 means weak correlation
## to be conservative, we choose |value| <0.01

df_sc_drop_corr = df_sc_drop.corr(method ='pearson').round(2) 
df_sc_drop_corr

# list the correlation between featurs and production was in the range of -0.2 to 0.2

rslt_corr = df_sc_drop_corr.loc[(df_sc_drop_corr['production']>=-0.2) & (df_sc_drop_corr['production']<=0.2)]

rslt_corr['production']

#################################

#############  Accroding to P value, we can delete 
######because (|p| value of  and  is larger than 0.05)

                # isip
                # pump rate

                
#############  Accroding to correlation value, we can delete

                # tvd (ft)
                # well spacing
                # log permeability
                # porosity 
                # poisson ratio
                # s-velocity
                # isip
                # breakdown pressure
                # pump rate
                # op_number
                
################################
# we will delete all attributes recommended by P value and VIF
# except, 'log permeability' and 'porosity' as the reference has reported the correlation to production 

############################## drop atrributes

df_sc_drop_2 = df_sc_drop.drop([#'tvd (ft)',
                                 #'well spacing',
                                 #'poisson ratio',
                                 #'s-velocity',
                                 'isip',
                                 'breakdown pressure',
                                 'pump rate',
                                 #'op_number'
                                    ], axis=1)

###############################  again, use remaining varables to predict, find adj R2

x = df_sc_drop_2.drop(['production'], axis=1)
y = df_sc_drop_2['production']
                
X2 = sm.add_constant(x)
est = sm.OLS(y, X2)
est2 = est.fit()
print(est2.summary())

######################
# as observed delete three attributes, the adj R2 does not significantly change
# the p value <0.05, the prediction model so far is well-built

################################################# again, use remaining varables to predict, find adj R2

############################## drop atrributes

df_sc_drop_3 = df_sc_drop_2.drop([#'tvd (ft)',
                                 #'well spacing',
                                #'log permeability',
                                #'porosity',
                                #'poisson ratio',
                                #'p-velocity',
                                #'s-velocity',
                                #'youngs modulus',
                                 #'total number of stages',
                                #'proppant volume',
                                 #'date_sec'#,
                                 #'op_number'
                                    ], axis=1)

###############################  again, use remaining varables to predict, find adj R2

x = df_sc_drop_3.drop(['production'], axis=1)
y = df_sc_drop_3['production']
                
X2 = sm.add_constant(x)
est = sm.OLS(y, X2)
est2 = est.fit()
print(est2.summary())
#### result, every variable is important

#### result, every variable is important

## record the attributes and use the orginal data in the attributes to repeat the adj R 2

df_att = df[['tvd (ft)',
             'well spacing',
             'log permeability',
             'porosity',
             'poisson ratio',
             'p-velocity',
             's-velocity', 
             'youngs modulus', 
             'total number of stages',
             'proppant volume',
             'date_sec',
              'op_number',
             'production']]
df_att_drop = df_att.dropna()

###############################  again, use remaining varables to predict, find adj R2

x = df_att_drop.drop(['production'], axis=1)
y = df_att_drop['production']
                
X2 = sm.add_constant(x)
est = sm.OLS(y, X2)
est2 = est.fit()
print(est2.summary())
#### result, every variable is important

################################# correlation ##########################################

## the coefficient returns a value between -1 and 1 
## the value represents the limits of correlation from a full negative correlation to a full positive correlation
## according to reference, |value| <0.1 means weak correlation
## to be conservative, we choose |value| <0.01

df_att_drop_corr = df_att_drop.corr(method ='pearson').round(2) 
df_att_drop_corr

# list the correlation between featurs and production was in the range of -0.2 to 0.2

################################# confirm the critical variables ##########################################
## after check with the original dataset
## we confirm that the critical variables are 
                                            #   'tvd (ft)', 
                                            #   'well spacing',
                                            #   'log permeability',            
                                            #   'porosity',            
                                            #   'poisson ratio',           
                                            #   'p-velocity',          
                                            #   's-velocity',             
                                            #   'youngs modulus',             
                                            #   'total number of stages',           
                                            #   'proppant volume',           
                                            #   'date_sec',          
                                            #   'op_number'
