==> step4_improve model using fix effects.ipynb <==
############################################################################################################################
#############################                     convert data into df                    ##################################
############################################################################################################################


import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
from patsy import dmatrices
import statsmodels.api as sm


from statsmodels.stats.outliers_influence import variance_inflation_factor
from statsmodels.tools.tools import add_constant


# read data as framework
df = pd.read_csv('./data_integer_shift_column.csv')

# check the framework

#df.head()

############################## according to step 2 the following attributes are important for regression

                                #'tvd (ft)',
                                # 'well spacing',
                                # 'log permeability',
                                # 'porosity',
                                # 'poisson ratio',
                                # 'p-velocity',
                                # 's-velocity', 
                                # 'youngs modulus', 
                                # 'total number of stages',
                                # 'proppant volume',
                                # 'date_sec',
                                # 'op_number',
                                # 'production'

############################## according to step 3 adj R2 was too small ~0.55


############################## thus, the model need to be improved


############################################################################################################################
###################          use fix effect to improve the model and increase adj R2           ##########################################
######################################################################################################################



##################### record the attributes and use the orginal data in the attributes to repeat the adj R 2

df_att = df[['tvd (ft)',
             'well spacing',
             'log permeability',
             'porosity',
             'poisson ratio',
             'p-velocity',
             's-velocity', 
             'youngs modulus', 
             'total number of stages',
             'proppant volume',
             'date_sec',
              'op_number',
             'production']]
df_att_drop = df_att.dropna()


##################### set x and y for the models
x = df_att_drop.drop(['production'], axis=1)
y = df_att_drop['production']

###############################  use critical varables to predict, find adj R2 BEFORE 'fixed effect' function
X2 = sm.add_constant(x)
est = sm.OLS(y, X2)
est2 = est.fit()
print(est2.summary())

#### The adj R2 before improvement was 0.552

###############################  use critical varables to predict, find adj R2 before 'fixed effect' function

df_FE = sm.OLS(endog=y, exog=x, entity_effects=True).fit()
print(df_FE.summary())

# Another syntex for Fix effects


# FE_ols = sm.OLS(endog=y, exog=x, formula='logrealwage ~ HHI+C(Cz)+C(yQ)-1',
#                               data=df_att_drop).fit()
# print(FE_ols.summary())

#########################################     find the attribute with too large P value (P >0.1)

### according to reference, if |P value| < 0.05, the feature is significant
### to be conservative, we choose |P value| < 0.1, and delete the value =>0.1


### so we need to delete 'operator'

#########################################     delete 'operator'

df_att_drop_2 = df_att_drop.drop(['op_number'], axis=1)

##################### set x and y for the models
x = df_att_drop_2.drop(['production'], axis=1)
y = df_att_drop_2['production']

###############################  use critical varables to predict, find adj R2 before 'fixed effect' function

df_FE = sm.OLS(endog=y, exog=x, entity_effects=True).fit()
print(df_FE.summary())

#### after deleting 'op_number', i.e., 'operator', the changes in R2 was negligible
#### this indicates that our deletion was reasonable -  does not hurt the model

############################################################################################################################
###############################                     find coefficient            ############################################
############################################################################################################################


###############################          covert summary table into DataFrame  
rslt_sum = df_FE.summary().tables[1]
df_rslt_sum = pd.DataFrame(rslt_sum)
df_rslt_sum

###############################          use the first raw as the columns name  

#grab the first row for the header
new_header = df_rslt_sum.iloc[0] 

#take the data less the header row
df_rslt_sum = df_rslt_sum[1:] 

#set the header row as the df header
df_rslt_sum.columns = new_header

df_rslt_sum

###############################          use the first column as the index

#rename the first column
df_rslt_sum_new = df_rslt_sum.rename(columns={ df_rslt_sum.columns[0]: 'features' })

df_rslt_sum_new

# list the data type
list(df_rslt_sum_new.columns) 

# convert DataFrame as CSV file and then convert back to change the 'statsmodels.iolib.table.Cell' format

df_rslt_sum_new.to_csv('df_transition.csv',index=False)

# read data as framework
df_transition = pd.read_csv('./df_transition.csv')

############################################################################################################################
###############################                     result            ############################################
############################################################################################################################

## R2 was 0.847
## adj R2 was 0.844

# read data as framework
df_transition[['features','coef','P>|t|']]
